dropdown menu modified to be an overlay of #playlists instead of below #title-search (for ease of sizing). rewired dropdown menu coverage, may require some more work though
