+++ frontend/index.html
@@ -81,19 +81,17 @@
                         <i class="fa fa-rocket"></i>
                     </button>
                 </div>
-
-                
-                <div id="search-dropdown" class="dropdown hidden">
-                    <ul id="live-search-results" class="list-track">
-                        <li>search 1</li>
-                        <li>search 2</li>
-                        <li>search 3</li>
-                    </ul>
-                </div>
                 
             </div>
 
             <div id="playlists">
+                <!-- search dropdown menu absolutely positioned inside the playlists body -->
+                <div id="search-dropdown" class="hidden">
+                    <ul id="live-search-results" class="list-track">
+                    </ul>
+                </div>
+
+
                 <div id="library" class="playlist">
                     <div class="list-header">
                         <h3 class="list-title">Library</h3>
