+++ frontend/js/events/dom/toggle.js
@@ -1,5 +1,7 @@
 import { isMobile } from "../../utils/index.js";
 
+import { collapsedHeight } from "../../dom/index.js";
+
 let isCollapsed = true;
 
 const toggleButton = document.getElementById("queue-toggle-button");
@@ -7,29 +9,6 @@ const container = document.getElementById("playbar-queue");
 const playbar = document.getElementById("playbar");
 const toast = document.getElementById("toast");
 const marginBlock = document.getElementById("title-search-playlists-margin-block");
-
-//do this once to save the collapsed height
-function measureCollapsedHeight() {
-    if (isMobile()) {
-        //mobile collapsed height measurement
-        const rect = playbar.getBoundingClientRect();
-        const styles = getComputedStyle(playbar);
-        const marginTop = parseFloat(styles.marginTop) || 0;
-        return (rect.height + marginTop);
-
-        // console.log("test", rect.height, marginTop);
-        // console.log(window.getComputedStyle(playbar).height);
-    } else {
-        //desktop collapsed height measurement including button
-        const rect = playbar.getBoundingClientRect();
-        const butt = toggleButton.getBoundingClientRect();
-        const styles = getComputedStyle(playbar);
-        const marginTop = parseFloat(styles.marginTop) || 0;
-        const marginBottom = parseFloat(styles.marginBottom) || 0;
-        return (rect.height + butt.height + marginTop + marginBottom);
-    }
-}
-const collapsedHeight = measureCollapsedHeight();
 
 
 //set other elements' margin block
